+++ lib/src/data/streaming/streaming_request.dart
@@ -1,21 +1,19 @@
-
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:misskey_dart/src/enums/channel.dart';
-import 'package:misskey_dart/src/enums/channel_data_type.dart';
+import 'package:misskey_dart/src/enums/streaming_request_type.dart';
 
-part 'channel_request.freezed.dart';
-part 'channel_request.g.dart';
+part 'streaming_request.freezed.dart';
+part 'streaming_request.g.dart';
 
 @freezed
 class ChannelRequest with _$ChannelRequest {
   const factory ChannelRequest({
-    @ChannelDataTypeJsonConverter()
-    required ChannelDataType type,
+    @ChannelDataTypeJsonConverter() required ChannelDataType type,
     required ChannelRequestBody body,
   }) = _ChannelRequest;
 
-  factory ChannelRequest.fromJson(Map<String, Object?> json)
-      => _$ChannelRequestFromJson(json);
+  factory ChannelRequest.fromJson(Map<String, Object?> json) =>
+      _$ChannelRequestFromJson(json);
 }
 
 @freezed
@@ -26,6 +24,6 @@ class ChannelRequestBody with _$ChannelRequestBody {
     required Map<String, dynamic>? params,
   }) = _ChannelRequestBody;
 
-  factory ChannelRequestBody.fromJson(Map<String, Object?> json)
-      => _$ChannelRequestBodyFromJson(json);
+  factory ChannelRequestBody.fromJson(Map<String, Object?> json) =>
+      _$ChannelRequestBodyFromJson(json);
 }
