+++ lib/src/data/streaming/streaming_response.dart
@@ -1,0 +1,30 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:misskey_dart/src/enums/channel.dart';
+import 'package:misskey_dart/src/enums/streaming_request_type.dart';
+import 'package:misskey_dart/src/enums/channel_event_type.dart';
+
+part 'streaming_response.freezed.dart';
+part 'streaming_response.g.dart';
+
+@freezed
+class ChannelResponse with _$ChannelResponse {
+  const factory ChannelResponse({
+    @ChannelDataTypeJsonConverter() required ChannelDataType type,
+    required ChannelResponseBody body,
+  }) = _ChannelResponse;
+
+  factory ChannelResponse.fromJson(Map<String, Object?> json) =>
+      _$ChannelResponseFromJson(json);
+}
+
+@freezed
+class ChannelResponseBody with _$ChannelResponseBody {
+  const factory ChannelResponseBody({
+    required String id,
+    @ChannelResponseTypeJsonConverter() required ChannelResponseType type,
+    required Map<String, dynamic>? body,
+  }) = _ChannelResponseBody;
+
+  factory ChannelResponseBody.fromJson(Map<String, Object?> json) =>
+      _$ChannelResponseBodyFromJson(json);
+}
