+++ lib/src/enums/channel_event_type.dart
