+++ lib/src/enums/streaming_request_type.dart
